backend/requirements.txt
fastapi==0.115.0; python_version < "3.14"
fastapi==0.135.2; python_version >= "3.14"
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.2; python_version < "3.14"
pydantic==2.12.5; python_version >= "3.14"
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
psycopg2-binary==2.9.10; python_version < "3.14"
psycopg[binary]==3.2.13; python_version >= "3.14"
python-dotenv==1.0.1
